Another BSD compile fix

+++ lib/secure/src/olsrd_secure.c
@@ -351,7 +351,7 @@ add_signature(uint8_t *pck, int *size)
   
   /* Add timestamp */
   msg->sig.timestamp = htonl(now.tv_sec);
-  olsr_printf(3, "[ENC]timestamp: %ld\n", now.tv_sec);
+  olsr_printf(3, "[ENC]timestamp: %lld\n", (long long)now.tv_sec);
   
   /* Set the new size */
   *size += sizeof(struct s_olsrmsg);
@@ -518,7 +518,7 @@ validate_packet(struct interface *olsr_if, const char *pck, int *size)
       return 0;
     }
 
-  olsr_printf(1, "[ENC]Received timestamp %ld diff: %ld\n", rec_time, now.tv_sec - rec_time);
+  olsr_printf(1, "[ENC]Received timestamp %lld diff: %lld\n", (long long)rec_time, (long long)now.tv_sec - (long long)rec_time);
 
   /* Remove signature message */
   *size = packetsize;
@@ -986,7 +986,7 @@ send_cres(struct interface *olsr_if, union olsr_ip_addr *to, union olsr_ip_addr 
 
   /* set timestamp */
   crmsg.timestamp = now.tv_sec;
-  olsr_printf(3, "[ENC]Timestamp %ld\n", crmsg.timestamp);
+  olsr_printf(3, "[ENC]Timestamp %lld\n", (long long)crmsg.timestamp);
 
   /* Fill subheader */
   memcpy(&crmsg.destination, to, olsr_cnf->ipsize);
@@ -1066,7 +1066,7 @@ send_rres(struct interface *olsr_if, union olsr_ip_addr *to, union olsr_ip_addr 
 
   /* set timestamp */
   rrmsg.timestamp = now.tv_sec;
-  olsr_printf(3, "[ENC]Timestamp %ld\n", rrmsg.timestamp);
+  olsr_printf(3, "[ENC]Timestamp %lld\n", (long long)rrmsg.timestamp);
 
   /* Fill subheader */
   memcpy(&rrmsg.destination, to, olsr_cnf->ipsize);
